refactor: extract browser-runtime-core
