+++ packages/browser-runtime-core/.eslintignore
@@ -1,0 +1,4 @@
+/lib/**
+/build/**
+/coverage/**
+!.eslintrc.js
