+++ packages/browser-runtime-core/src/interpreter/interpreter.ts
