+++ packages/browser-runtime-core/src/interpreter/preprocessor/collect-top-level-lexical-context.ts
