+++ packages/browser-runtime-core/src/interpreter/preprocessor/inject-last-expression-callback.spec.ts
@@ -1,4 +1,4 @@
-import { expect } from '../../../testing/chai';
+import { expect } from 'chai';
 
 import { parse } from '@babel/parser';
 import generate from '@babel/generator';
