+++ packages/browser-runtime-core/src/interpreter/preprocessor/save-and-restore-lexical-context.ts
