+++ packages/browser-runtime-core/src/interpreter/preprocessor/transform-command-invocation.ts
