+++ packages/browser-runtime-core/src/interpreter/preprocessor/wrap-object-literal.ts
