+++ packages/browser-runtime-core/src/open-context-runtime.ts
@@ -1,8 +1,8 @@
-import { Completion } from '../autocompleter/autocompleter';
+import { Completion } from './autocompleter/autocompleter';
 import { ServiceProvider } from 'mongosh-service-provider-core';
-import { ShellApiAutocompleter } from '../autocompleter/shell-api-autocompleter';
-import { Interpreter, InterpreterEnvironment, EvaluationResult } from '../interpreter';
-import { Runtime } from '../components/runtime';
+import { ShellApiAutocompleter } from './autocompleter/shell-api-autocompleter';
+import { Interpreter, InterpreterEnvironment, EvaluationResult } from './interpreter';
+import { Runtime } from './runtime';
 
 import Mapper from 'mongosh-mapper';
 import ShellApi from 'mongosh-shell-api';
