+++ packages/browser-runtime-core/src/runtime.ts
@@ -1,4 +1,4 @@
-import { Completion } from '../autocompleter/autocompleter';
+import { Completion } from './autocompleter/autocompleter';
 
 export type ContextValue = any;
